Starting iOSPorts Library Acknowledgment methods.
This library will provide a methods for including
acknowledgment for ports used by developers in
iOS apps.  Initially the library will allow 
developers to retrieve license, version, and package
information in the form of NSStrings.  Eventually
this will include UIViews and UIViewControllers
allowing for an easy drop to an application's
flow for port acknowledgments.